% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_emrgency.R
\docType{data}
\name{coffee}
\alias{coffee}
\title{Coffee Data Set}
\format{
A data frame with 4 rows and 2 variables:
\describe{
\item{time_stamp}{YYYY-MM-DD HH:MM:SS}
\item{amount}{Amount spent on coffee}
}
}
\usage{
coffee
}
\description{
Made-up data set for demonstrating \code{padr}.
}
\keyword{datasets}
