% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_emrgency.R
\docType{data}
\name{emergency}
\alias{emergency}
\title{Emergency Calls for Montgomery County, PA}
\format{
A data frame with 120450 rows and 6 variables:
\describe{
\item{lat}{Latitude from Google maps, based on the address}
\item{lng}{Longitude from Google maps, based on the address}
\item{zip}{Zipcode from Google, when possible}
\item{title}{Title, emergency category}
\item{time_stamp}{YYYY-MM-DD HH:MM:SS}
\item{twp}{Township}
}
}
\source{
\url{https://storage.googleapis.com/montco-stats/tzr.csv}
}
\usage{
emergency
}
\description{
The emergency calls coming in at Montgomery County, PA since 2015-12-10.
Data set was created at 2016-10-17 16:15:40 CEST from the API and contains
events until 2016-10-17 09:47:03 EST. From the original set the columns
desc and e are not included.
}
\keyword{datasets}
