% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_start_end.R
\name{format_start_end}
\alias{format_start_end}
\title{Add a Character of the Interval}
\usage{
format_start_end(x, start_format = "\%Y-\%m-\%d", end_format = start_format,
  sep = " ", end_is_start = TRUE, final_value = get_interval(x))
}
\arguments{
\item{x}{A vector of class \code{Date}, class \code{POSIXct} or class \code{POSIXlt}.}
}
\value{
`x` with a character column added.
}
\description{
Format a date variable to a character string of the interval in which the
events toke place.
}
\details{
The end of the periods will be determined by the next unique value
in `x`. It does so without regarding the interval of `x`. If a specific
interval is desired, `thicken` and / or `pad` should first be applied.
}
\examples{

}

