% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_helpers.R
\name{get_pulse}
\alias{get_pulse}
\title{Get the pulse of a datetime variable}
\usage{
get_pulse(x)
}
\arguments{
\item{x}{A variable of class of class \code{Date} or of class \code{POSIXt}.}
}
\value{
A character string indicating the pulse of \code{x}.
}
\description{
The pulse is the lowest highest time unit that can explain all instances of a
variable of class \code{Date} or of class \code{POSIXct}. This function
will determine what the pulse of the variable is.
}
\examples{
x_month <- seq(as.Date('2016-01-01'), as.Date('2016-05-01'), by = 'month')
get_pulse(x_month)
}

