% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_date.R
\name{interval_format_date}
\alias{interval_format_date}
\title{From Datetime to Interval character}
\usage{
interval_format_date(x, start_format = "\%Y-\%m-\%d",
  end_format = start_format, sep = " ", same_day = TRUE, colname = NULL,
  by = NULL, check_completeness = TRUE)
}
\description{
Format a date variable to a character string of the interval in which the
events toke place.
}

