% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/span_functions.R
\name{offset_date}
\alias{offset_date}
\title{Adjust date by offset}
\usage{
offset_date(x, date_offset)
}
\arguments{
\item{x}{A vector of class \code{Date}, \code{POSIXlt}, or \code{POSIXct}.}

\item{date_offset}{A numeric vector of respectively the year, month, day, hour,
minute and second to change in \code{x}.}
}
\value{
\code{x} with the six time levels specified in \code{offset} added
to it.
}
\description{
This function is a little helper to the span_ functions.
}
\examples{
 a_time_value <- as.POSIXct(strftime('2014-03-04 10:43:16'))
 desired_offset <- c(0, 2, 1, 2, 0, 18)
 offset_date(a_time_value, desired_offset)
}

