% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_df.R
\name{pad_date}
\alias{pad_date}
\title{Pad a data.frame}
\usage{
pad_date(df, pad = NULL, interval = NULL, with = c("0_or_NA", "NA",
  "last_caried_forward"), pad_present_NA = FALSE, ....)
}
\description{
This function will take a \code{data.frame} that contains a variable of the
following classes \code{Date}, \code{POSIXct}, or \code{POSIXlt}. It will
insert rows for all the instances that are missing from the data.
}
\details{
If the interval is not provided \code{pad_date} will determine the
interval itself by using the \code{get_interval} function. If the interval
is provided, but the interval present in the data is of a finer granualarity
it will automatically set the data to the closest instance of \code{interval}.
If multiple records should be set to
}

