% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padr_summarise.R
\name{padr_summarise}
\alias{padr_summarise}
\title{Summarise a Datetime Variable}
\usage{
padr_summarise(x, group = NULL, by = NULL)
}
\arguments{
\item{x}{A data frame containing at least one variable of class \code{Date},
class \code{POSIXct} or class \code{POSIXlt}.}

\item{group}{Optional character vector that specifies the grouping
variable(s).}

\item{by}{Only needs to be specified when \code{x} contains multiple
variables of class \code{Date}, class \code{POSIXct} or
class \code{POSIXlt}. \code{by} indicates which variable to summarise.}
}
\description{
For each group reports the first and the last observations, the interval, and
wether the time series is complete (if not it needs padding).
}
