% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thicken_helpers.R
\name{round_thicken}
\alias{round_thicken}
\title{Find closest value of a in b}
\usage{
round_thicken(a, b, direction = c("down", "up"))
}
\arguments{
\item{a}{A vector of class \code{Date}, \code{POSIXct}, or \code{POSIXlt}.}

\item{b}{A vector of the same class as \code{a}.}
}
\value{
A vector of the same class and the same length as \code{a}, with
for each value of \code{a} the closest value in \code{b}, either rounded up
or rounded down.
}
\description{
For two vectors of class \code{Date}, \code{POSIXct}, or \code{POSIXlt} find
the closest value of \code{a} in \code{b} under the restriction that the
returned value is either smaller than the value in \code{a}, when using
\code{round_down}, or larger than the value in \code{b}.
}

