% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/span_date.R
\name{span_date}
\alias{span_date}
\title{Wrapper around `seq.Date`}
\usage{
span_date(from, to = NULL, len_out = NULL, interval = NULL)
}
\arguments{
\item{from}{Integer of length 4 (year only), 6 (year and month), or 8 (
year, month, and day). Indicating the start value of the sequence.}

\item{to}{Integer of the same length as `from`.
Indicating the end value of the sequence. Optional.}

\item{len_out}{The desired length of the sequence. Optional.}

\item{interval}{The desired interval. Optional.}
}
\value{
An object of class Date.
}
\description{
Quickly create a sequence of dates from integer values.
}
\details{
In addition to `from`, `to` or `length` must be specified. If `from`
length is 4, the first day of that year will be used and the interval is
year. If its length is 6, the first day of that month will be used,
and the interval is month. If the length is 8, that day will be used and the
interval is day. Use `interval` to override these defaults.
}
\examples{
# using to
span_date(2011, 2015)
span_date(201101, 201501)
span_date(2011, 2015, interval = "month")
span_date(20110101, 20110201)
span_date(20110101, 20110201, interval = "month")

# using len_out
span_date(2011, len_out = 4)
span_date(201101, len_out = 4)
span_date(20110101, len_out = 4)
span_date(20110101, len_out = 4, interval = "month")
}

