% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/span_date.R
\name{span_time}
\alias{span_time}
\title{Wrapper around `seq.POSIXct`}
\usage{
span_time(from, to = NULL, len_out = NULL, interval = NULL, tz = "UTC")
}
\arguments{
\item{from}{Integer or character of length 4 (yyyy), 6 (yyyymm), or 8 (
yyymmdd). Character of length 11 (yyyymmdd hh), 13 (yyyymmdd hhmm), or 15 (
yyyymmdd hhmmss). Indicating the start value of the sequence.}

\item{to}{Integer or character of length 4 (yyyy), 6 (yyyymm), or 8 (
yyymmdd). Character of length 11 (yyyymmdd hh), 13 (yyyymmdd hhmm), or 15 (
yyyymmdd hhmmss). Indicating the end value of the sequence. Optional.}

\item{len_out}{The desired length of the sequence. Optional.}

\item{interval}{The desired interval. Optional.}

\item{tz}{The desired timezone.}
}
\value{
An object of class POSIXct.
}
\description{
Quickly create a sequence of dates from minimal specifications of datetimes.
}
\details{
Minimal specifications of datetimes set unspecified date parts to
default values. These are 01 for both month and day and 00 for hour, minute,
and second.

In addition to `from`, `to` or `length` must be specified. If the
interval is not specified, `span_time` will set the interval to the highest
of the specified datetime parts in either `from` or `to`. E.g. if they are
"20160103 01" and "20160108 05" it will be "day", if they are "2011" and
"20110101 021823" it will be "second".
}
\examples{
# using to
span_time(2011, 2013)
span_time("2011", "2013")
span_time(2011, 201301)
span_time(2011, 20130101)
span_time(2011, "20110101 0023")
span_time(2011, "20110101 002300")

# using len_out
span_time(2011, len_out = 3)
span_time("2011", len_out = 3)
span_time(2011, len_out = 10, interval = "month")
span_time(2011, len_out = 10, interval = "day")
span_time(2011, len_out = 10, interval = "hour")
span_time("20110101 00", len_out = 10)
span_time("20110101 002300", len_out = 10)

}

