% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thickr_helpers.R
\name{span_year}
\alias{span_day}
\alias{span_hour}
\alias{span_minute}
\alias{span_month}
\alias{span_year}
\title{Span a vector of dates}
\usage{
span_year(x)

span_month(x)

span_day(x)

span_hour(x)

span_minute(x)
}
\arguments{
\item{x}{A vector of class \code{Date}, \code{POSIXlt}, or \code{POSIXct} for
\code{span_year} or \code{span_month}. A vector of class  \code{POSIXlt}, or
\code{POSIXct} for the others.}
}
\value{
A vector of the same data type as \code{x}.
}
\description{
The spanning functions take a vector of class \code{Date}, \code{POSIXlt}, or
\code{POSIXct} and span a vector of its given interval from the first instance
smaller than the minimum of \code{x} to the first instance larger than the
maximum of \code{x}. The functions are designed as hepers for the \code{thickr}
function.
}
\examples{
x <- as.POSIXct(strftime(c('2014-03-04 10:43:16',
                           '2014-03-05 08:22:12')))
span_year(x)
span_month(x)
span_day(x)
span_hour(x)
span_minute(x)
}

