% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_to_middle.R
\name{center_interval}
\alias{center_interval}
\title{Shift to the Middle of the Interval}
\usage{
center_interval(x)
}
\arguments{
\item{x}{A vector of class \code{Date}, class \code{POSIXct} or class \code{POSIXlt}.}

\item{shift}{Up or down.}

\item{return}{`x` with the values shifted to the (approximate) center.}
}
\description{
After padding and thickening all the values are either
shifted to the first or the last value of the interval.
Function creates a vector that shifts the datetime value to
the (approximate) center of the interval.
}
\details{
The interval will be translated to number of days when
x is of class `Date``, or number of seconds when x is of class
`POSIXt`. For months and quarters will this be the average
length of the period. The floor of the translated number divided by two
and added by or subtracted from `x`.
}
\examples{
library(tidyverse)
plot_set <- emergency \%>\%
  thicken("hour", "h") \%>\%
  count(h) \%>\%
  head(24)

ggplot(plot_set, aes(h, n)) + geom_col()

plot_set \%>\%
  mutate(h_center = center_interval(h)) \%>\%
  ggplot(plot_set, aes(h_center, n)) + geom_col()
}

