% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_funcs.R
\name{format_start_end}
\alias{format_start_end}
\title{Make a Period Character vector}
\usage{
format_start_end(x, start_format = "\%Y-\%m-\%d", end_format = start_format,
  sep = " ", end_offset = 0, units_to_last = NULL)
}
\arguments{
\item{x}{A vector of class \code{Date}, class \code{POSIXct} or class \code{POSIXlt},
of which the values are unique.}
}
\value{
A character vector showing the interval.
}
\description{
After applying `thicken` all the observations in an interval are mapped
to a single timepoint. This function will conver a datetime variable to
a character vector that reflects the full period.
}
\details{
The end of the periods will be determined by the next unique value
in `x`. It does so without regarding the interval of `x`. If a specific
interval is desired, `thicken` and / or `pad` should first be applied to
create an equally spaced datetime variable.
}
\examples{
library(dplyr)
library(ggplot2)
plot_set <- emergency \%>\%
  head(500) \%>\%
  thicken("hour", "h") \%>\%
  count(h)

# this will show the data on the full hour
ggplot(plot_set, aes(h, n)) + geom_col()

# adding a character to indicate the hours of the interval.
plot_set \%>\%
  mutate(h_int = format_start_end(h, "\%H", sep = "-"))
}
