% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_df_helpers.R
\name{get_interval}
\alias{get_interval}
\title{Determine the interval of a date variable.}
\usage{
get_interval(x)
}
\arguments{
\item{x}{A variable of class of class \code{Date} or of class \code{POSIXt}.}
}
\value{
A character string indicating the interval of \code{x}.
}
\description{
The interval is the lowest time level that can explain all instances of a
variable of class \code{Date} or of class \code{POSIXct}. This function
will determine what the interval of the variable is. It is used within
the \code{pad_date} function when the interval is not provided.
}
\examples{
x_month <- seq(as.Date('2016-01-01'), as.Date('2016-05-01'), by = 'month')
get_interval(x_month)
}

