% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_week_start.R
\name{get_week_start}
\alias{get_week_start}
\title{Retrieve the closest given Weekday}
\usage{
get_week_start(wday = 2)
}
\arguments{
\item{wday}{Integer in the range 1-7 specifying the desired weekday start
(1 = Sun, 2 = Mon, 3 = Tue, 4 = Wed, 5 = Thu, 6 = Fri, 7 = Sat).}
}
\value{
Object of class "weekstart", only useful within `thicken`.
}
\description{
When applying \code{thicken} with the interval "week", use this function to
automatically start at the first weekday before \code{min(datetime_var)} at
`start_val` argument.
}
\examples{
library(dplyr)
coffee \%>\% thicken("week", start_val = get_week_start())
coffee \%>\% thicken("week", start_val = get_week_start(3))
}
