% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_cust.R
\name{pad_cust}
\alias{pad_cust}
\title{Apply pad with a custom spanning}
\usage{
pad_cust(x, spanned, by = NULL, group = NULL, drop_last_spanned = TRUE)
}
\arguments{
\item{x}{A data frame containing at least one datetime variable of
class \code{Date}, class \code{POSIXct} or class \code{POSIXlt}.}

\item{spanned}{A datetime vector to which the the datetime variable in `x`
should be mapped. See `subset_span` (TODO link) for quickly spanning unequally
spaced variables.}

\item{by}{Only needs to be specified when \code{x} contains multiple
variables of class \code{Date}, class \code{POSIXct} or class \code{POSIXlt}.}

\item{group}{Optional character vector that specifies the grouping
variable(s). Padding will take place within the different group values.}

\item{drop_last_spanned}{Logical, indicating wether to drop the last value
from `spanned`. The spanned is typically around the datetime variable. When
thickened down, this would create an empty record after the observations when
padding.
\code{by} indicates which to use for thickening.}
}
\value{
The data frame \code{x} with
}
\description{
Pad the datetime variable after `thicken_cust` is applied, using the same
spanning.
}
