% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_cust.R
\name{pad_cust}
\alias{pad_cust}
\title{Apply pad with a custom spanning}
\usage{
pad_cust(x, spanned, by = NULL, group = NULL, drop_last_span = TRUE)
}
\arguments{
\item{x}{A data frame containing at least one datetime variable of
class \code{Date}, class \code{POSIXct} or class \code{POSIXlt}.}

\item{spanned}{A datetime vector to which the the datetime variable in `x`
should be mapped. See `subset_span` (TODO link) for quickly spanning unequally
spaced variables.}

\item{by}{Only needs to be specified when \code{x} contains multiple
variables of class \code{Date}, class \code{POSIXct} or class \code{POSIXlt}.
\code{by} indicates which to use for thickening.}

\item{group}{Optional character vector that specifies the grouping
variable(s). Padding will take place within the different group values.}

\item{drop_last_span}{The `spanned` for `thicken_cust` needs to be around the
datetime variable. Typically, you don't want the last value in `spanned` for
padding, because it adds an empty record. When TRUE the last value in `spanned`
is removed before padding.}
}
\value{
The data frame \code{x} with
}
\description{
Pad the datetime variable after `thicken_cust` is applied, using the same
spanning.
}
