% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/span.R
\name{span}
\alias{span}
\title{Span a vector of dates}
\usage{
span(x, pulse = c("year", "quarter", "month", "week", "day", "hour", "min",
  "sec"), start_val = NULL)
}
\arguments{
\item{x}{A vector of class \code{Date}, \code{POSIXlt}, or \code{POSIXct} for
\code{span_year} or \code{span_month}. A vector of class  \code{POSIXlt}, or
\code{POSIXct} for the others.}

\item{start_val}{An object of class \code{Date}, \code{POSIXlt}, or \code{POSIXct}.
See \code{details} for further information.}

\item{end_va}{lAn object of class \code{Date}, \code{POSIXlt}, or \code{POSIXct}.
#' See \code{details} for further information.}
}
\value{
A vector of the same data type as \code{x}.
}
\description{
The spanning functions take a vector of class \code{Date}, \code{POSIXlt}, or
\code{POSIXct} and span a vector of its given interval from the first instance
smaller than the minimum of \code{x} to the first instance larger than the
maximum of \code{x}. The functions are designed as hepers for the \code{thicken}.
function.
}
\details{
The \code{start_val} and \code{end_val} arguments can be used in two ways to
change the default behavior of the \code{span_} functions. First one can
lengthen or shorten the output by specifying a timepoint that is outside
the default range. Second by default the lower order time units are set to the
lowest possible value (that is 1 for month and day and 0 for hour, minute and
second). \code{start} and \code{end} can be used as an offset for these
default values. Of course these two functionalities are not mutually exclusive.
}
\examples{
x <- as.POSIXct(strftime(c('2014-03-04 10:43:16',
                           '2014-03-05 08:22:12')))
span_year(x)
span_month(x)
span_day(x)
span_hour(x)
span_minute(x)
}

