% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thicken.R
\name{thicken}
\alias{thicken}
\title{Change datetime Granularity}
\usage{
thicken(x, interval = c("year", "month", "day", "hour", "minute"),
  rounding = c("closest", "up", "down"), allow_duplicates = TRUE, ...)
}
\arguments{
\item{x}{A vector of class \code{Date}, \code{POSIXlt}, or \code{POSIXct}.}

\item{interval}{The desired interval of the output, which should be higher
the interval of \codes{x}. Current choices are \code{year, month, day,
hour, minute}.}

\item{rounding}{= closest Should \code{x} be rounded to the \code{closest}
instance of \code{interval} or should it be rounde \code{up} or \code{down}?}

\item{allow_duplicates}{Logical indicating if we can have multiple instances
of the same value in the return \code{data.frame}.
If FALSE the value in \code{x} closest to the
return time point will be selected. Will throw an error if FALSE and two or
more timpe points are at the exact same distance of an \code{interval} value.}
}
\value{
A \code{data.frame} with two colums, the original vector \code{x}
and a vector of the same class as \code{x} with the datetime points to which
\code{x} should be mapped to thicken it.  Datetime levels that are
lower in granularity than
possible value (month = 1, day = 1, hour = 0, second = 0, minute = 0).
}
\description{
This function will take a datetime vector of class and changes its
interval level to the provided level.
}
\details{
The interval is moved to a higher level by 'rounding' \code{x}
to the closest instance of the provided \code{interval}
}
\examples{

}

