% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smear.R, R/thicken.R
\name{smear}
\alias{smear}
\alias{thicken}
\title{Change datetime interval up}
\usage{
smear(x, interval = c("year", "month", "day", "hour", "minute"),
  rounding = c("down", "up"), by = NULL, start_val = NULL,
  end_val = NULL)

thicken(x, interval = c("year", "month", "day", "hour", "minute"),
  rounding = c("down", "up"), by = NULL, start_val = NULL,
  end_val = NULL)
}
\arguments{
\item{x}{A vector of class \code{Date}, \code{POSIXlt}, or \code{POSIXct},
or a \code{data.frame} containing at least one column of these classes.}

\item{interval}{The desired interval of the output, which should be higher
then the interval of the datetime variable. Current choices are
\code{year, month, day, hour, minute}.}

\item{rounding}{= closest Should \code{x} be rounded \code{down} or \code{up}?}

\item{by}{If \code{x} is a \class{data.frame} and contains multiple datetime
variables, specify which column to pad by.}

\item{start_val}{Change the default start_valpoint of the time range to which
the datetime variable is thickened. See ?span_year for more information.}

\item{end_val}{Change the default end_valpoint of the time range to which
the datetime variable is thickened. See ?span_year for more information.}
}
\value{
The original \code{data.frame x} with the thickened variable added
to it. If \code{x} is a datetime vector the return will be a
\code{data.frame} comprising \code{x} and the thickened variable.
}
\description{
This function will take a datetime vector of class code{Date}, \code{POSIXlt}, or \code{POSIXct}
and changes its interval up level to the provided level.
}
\details{
The interval is moved to a higher level by 'rounding' \code{x}
to the closest instance of the provided \code{interval}
}
\examples{
X <- data.frame(day_var = seq(as.Date('2016-01-01'), as.Date('2016-12-31'), by = 'day'),
                value   = runif(366, 50, 100))
}

