% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thicken_cust.R
\name{thicken_cust}
\alias{thicken_cust}
\title{Apply thicken with a custom spanning}
\usage{
thicken_cust(x, spanned, colname, rounding = c("down", "up"), by = NULL)
}
\arguments{
\item{x}{A data frame containing at least one datetime variable of
class \code{Date}, class \code{POSIXct} or class \code{POSIXlt}.}

\item{spanned}{A datetime vector to which the the datetime variable in `x`
should be mapped. See `subset_span` (TODO link) for quickly spanning unequally
spaced variables.}

\item{colname}{The column name of the added variable.}

\item{rounding}{Should a value in the input datetime variable be mapped to
the closest value that is lower (\code{down}) or that is higher (\code{up})
than itself.}

\item{by}{Only needs to be specified when \code{x} contains multiple
variables of class \code{Date}, class \code{POSIXct} or class \code{POSIXlt}.
\code{by} indicates which to use for thickening.}
}
\value{
The data frame \code{x} with the variable added to it.
}
\description{
Like `thicken` this function will find the datetime variable in `x`, and add
a variable of a higher periodicity to it. However, the variable to which to
map the observation is provided by the user. This enables mapping to
time points that are unequally spaced.
}
\examples{
library(dplyr)
# analysis of traffic accidents in traffic jam hours and other hours.
accidents <- emergency \%>\% filter(title == "Traffic: VEHICLE ACCIDENT -")
spanning <- span_time("20151210 16", "20161017 17", tz = "EST") \%>\%
  subset_span(list(hour = c(6, 9, 16, 19)))
thicken_cust(accidents, spanning, "period") \%>\%
  count(period)
}
